+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Extensible formatters for table content in content element (#10)
+
 ## [0.14.0] - 2021-11-21
 
 ### Added