+++ Documentation/changelog.rst
@@ -11,6 +11,12 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 `Unreleased <https://github.com/brotkrueml/typo3-jobrouter-data/compare/v0.14.0...HEAD>`_
 ---------------------------------------------------------------------------------------------
 
+Added
+^^^^^
+
+
+* Extensible formatters for table content in content element (#10)
+
 `0.14.0 <https://github.com/brotkrueml/typo3-jobrouter-data/compare/v0.13.0...v0.14.0>`_ - 2021-11-21
 ---------------------------------------------------------------------------------------------------------